+++ lib/main.dart
@@ -8,7 +8,7 @@ Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await LiquidGlassWidgets.initialize();
 
-  ms.applicationVersion = "0.1.0+0621261707";
+  ms.applicationVersion = "0.2.0+0623260007";
 
   runApp(LiquidGlassWidgets.wrap(child: Switchboard()));
 }